requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-allauth==0.48.0
django-autoslug==1.9.8
django-bootstrap4==21.2
django-crispy-forms==1.14.0
django-model-utils==4.2.0
django-utils==0.0.2
django-ckeditor==.3.0
idna==3.3
mod-wsgi==4.9.0
oauthlib==3.2.0
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
utils==1.0.1
django-forms-dynamic==1.0.0
django-widget-tweaks==1.4.12


